+++ src/pretalx/api/serializers/question.py
@@ -38,11 +38,23 @@ class MinimalQuestionSerializer(ModelSerializer):
         fields = ("id", "question")
 
 
-class AnswerSerializer(ModelSerializer):
-    question = MinimalQuestionSerializer(Question.objects.none(), read_only=True)
-    submission = SlugRelatedField(queryset=Submission.objects.none(), slug_field="code")
-    person = SlugRelatedField(queryset=User.objects.none(), slug_field="code")
-    options = AnswerOptionSerializer(many=True)
+class AnswerWriteSerializer(ModelSerializer):
+    submission = SlugRelatedField(
+        queryset=Submission.objects.none(), slug_field="code", required=False
+    )
+    person = SlugRelatedField(
+        queryset=User.objects.all(), slug_field="code", required=False
+    )
+    options = AnswerOptionSerializer(many=True, required=False)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        request = self.context.get("request")
+        if not request:
+            return
+        self.fields["question"].queryset = request.event.questions.all()
+        self.fields["submission"].queryset = request.event.submissions.all()
+        self.fields["review"].queryset = request.event.reviews.all()
 
     class Meta:
         model = Answer
@@ -52,6 +64,14 @@ class AnswerSerializer(ModelSerializer):
             "answer",
             "answer_file",
             "submission",
+            "review",
             "person",
             "options",
         )
+
+
+class AnswerSerializer(AnswerWriteSerializer):
+    question = MinimalQuestionSerializer(Question.objects.none())
+
+    class Meta(AnswerWriteSerializer.Meta):
+        pass