+++ doc/changelog.rst
@@ -3,7 +3,7 @@
 Release Notes
 =============
 
-- :feature:`-` There is a new API endpoint, ``/questions/``, that incientally is our first writable API endpoint.
+- :feature:`-` There are two new API endpoints, ``/questions/`` and ``/answers/``, that incientally are our first writable API endpoints. The API docs have been updated.
 - :support:`-` Email error reporting (sent to instance administrators) now includes a short explanation and a link to the pretalx issue tracker.
 - :feature:`-` If a speaker has selected to show their gravatar, it is now also exposed in the API in the avatar field.
 - :feature:`-` When you send out reminders about unanswered questions, you can now target specific questions, or tracks, or submission types.